bonus/requirements-bonus-ttt.txt
tensorflow-cpu==2.16.1
matplotlib==3.9.0
numpy==1.26.4
